LayoutTests/imported/w3c/web-platform-tests/custom-elements/registries/Document-importNode-cross-document.window-expected.txt


PASS Cloning with global registry
PASS Cloning with explicit global registry
FAIL Cloning with scoped registry assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL Cloning including shadow tree with global registry assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL Cloning including shadow tree with explicit global registry assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS Cloning including shadow tree with scoped registry
PASS Cloning with global registry (null registry target)
PASS Cloning with explicit global registry (null registry target)
FAIL Cloning with scoped registry (null registry target) assert_equals: expected object "[object CustomElementRegistry]" but got null
PASS Cloning including shadow tree with global registry (null registry target)
PASS Cloning including shadow tree with explicit global registry (null registry target)
FAIL Cloning including shadow tree with scoped registry (null registry target) assert_equals: expected object "[object CustomElementRegistry]" but got null
